Extract LayerStarsModule
