+++ src/app/modules/controls/components/selector-star-magnitude/selector-star-magnitude.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { SupportedLayers } from '#core/models/layers/supported-layers';
 import { LayersVisibilityManagerService } from '#core/services/visibility/layers-visibility-manager.service';
-import { StarsVisibilityManagerService } from '#core/services/visibility/stars-visibility-manager.service';
+import { StarsVisibilityManagerService } from '#layer-stars/services/stars-visibility-manager.service';
 
 /**
  * Provides the UI with the controls allowing to select
