+++ src/app/modules/controls/components/selector-star-names/selector-star-names.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { SupportedLayers } from '#core/models/layers/supported-layers';
 import { NameSelectionType } from '#controls/models/name-selection-type';
 import { LayersVisibilityManagerService } from '#core/services/visibility/layers-visibility-manager.service';
-import { StarsVisibilityManagerService } from '#core/services/visibility/stars-visibility-manager.service';
+import { StarsVisibilityManagerService } from '#layer-stars/services/stars-visibility-manager.service';
 import { TextsVisibilityManagerService } from '#core/services/visibility/texts-visibility-manager.service';
 
 /**
