+++ src/app/modules/layer-stars/models/stars.ts
